// mobile/src/types/PaymentTypes.ts
export interface Payment {
  id: number;
  number: string;
  number_ref: string;
  adapter: string;
  expired_at: string;
  currency_code: string;
  total: number;
  index: number;
  status: number;
  created_at: string;
  updated_at: string;
  request_currency_code: string;
  request_amount: number;
  request_method: string;
  response_currency_code?: string;
  response_amount?: number;
  response_method?: string;
  response_status?: string;
  response_key?: string;
  response_url?: string;
  response_redirect_success?: string;
  response_redirect_expired?: string;
  response_redirect_failed?: string;
  charge_currency_code?: string;
  charge_amount?: number;
  charge_method?: string;
  charge_status?: string;
  charge_paid_at?: string;
  qr_content?: string;
  qr_url?: string;
  qr_expired_at?: string;
  qr_name?: string;
  va_number?: string;
  va_name?: string;
  va_expired_at?: string;
  va_channel?: string;
  card_type?: string;
  card_brand?: string;
  simulation?: string;
  step?: string;
}